Avoid rendering cycle in partials
We need to avoid self referential and non-self referential
use cases or things will overflow. This avoids the condition
by ensuring that when a partial renders itself, it removes itself
from the available lookups downstream. If lookup failures are
enabled, then the operation will fail. Otherwise, it will
be a silent error.