feat: Add custom funcs to allow invoking mutations
This adds an Observe specific token `~` to act like a section
and attempt a look in registered customize functions. It
has an open and close part, just like a section, and everything
within the section is passed to the function for rendering.

This was added to support an immediate need for url-encoding
only a single part of a document, but likely will be used for
additional formatting or encoding needs.

It's possible we could allow optional options/arguments
in the lexing, but we don't need it right now. We aren't
prevented from adding it in the future though.