fix: Do not HTML escape in fallback JSON mode
When a template specifies `{{{.}}}` or otherwise
falls back to the json.Marshal case, it needs to avoid
escaping &, <, and > characters. This is especially
important because it is common to have URLs with query
strings in object properties and you get a different behavior
from referencing the property vs the parent object.